48-bit I/O addresses

+++ src/fpga.h
@@ -112,7 +112,7 @@ struct Rom {
 };
 
 struct AXI_Write_State {
-  uint32_t awaddr;
+  uint64_t awaddr;
   uint16_t wdata_count;
   uint16_t wid;
   AXI_Write_State(uint32_t awaddr, uint16_t wdata_count, uint16_t wid)
